refactor(redisfailover): combine with redis provider
