+++ providers/redis/provider.go
@@ -43,8 +43,89 @@ func New(cfg Config) (*Provider, error) {
 	}
 
 	p := &Provider{
-		config: cfg,
-		db:     db,
+		keyprefix: cfg.KeyPrefix,
+		db:        db,
+	}
+
+	return p, nil
+}
+
+// NewFailover returns a new configured redis failover (sentinel) provider
+func NewFailover(cfg FailoverConfig) (*Provider, error) {
+	if cfg.MasterName == "" {
+		return nil, errConfigMasterNameEmpty
+	}
+
+	db := redis.NewFailoverClient(&redis.FailoverOptions{
+		MasterName:         cfg.MasterName,
+		SentinelAddrs:      cfg.SentinelAddrs,
+		SentinelPassword:   cfg.SentinelPassword,
+		RouteByLatency:     cfg.RouteByLatency,
+		RouteRandomly:      cfg.RouteRandomly,
+		SlaveOnly:          cfg.SlaveOnly,
+		Username:           cfg.Username,
+		Password:           cfg.Password,
+		DB:                 cfg.DB,
+		MaxRetries:         cfg.MaxRetries,
+		MinRetryBackoff:    cfg.MinRetryBackoff,
+		MaxRetryBackoff:    cfg.MaxRetryBackoff,
+		DialTimeout:        cfg.DialTimeout,
+		ReadTimeout:        cfg.ReadTimeout,
+		WriteTimeout:       cfg.WriteTimeout,
+		PoolSize:           cfg.PoolSize,
+		MinIdleConns:       cfg.MinIdleConns,
+		MaxConnAge:         cfg.MaxConnAge,
+		PoolTimeout:        cfg.PoolTimeout,
+		IdleTimeout:        cfg.IdleTimeout,
+		IdleCheckFrequency: cfg.IdleCheckFrequency,
+		TLSConfig:          cfg.TLSConfig,
+	})
+
+	if err := db.Ping(context.Background()).Err(); err != nil {
+		return nil, errRedisConnection(err)
+	}
+
+	p := &Provider{
+		keyprefix: cfg.KeyPrefix,
+		db:        db,
+	}
+
+	return p, nil
+}
+
+// NewFailover returns a new configured redis cluster provider
+func NewCluster(cfg ClusterConfig) (*Provider, error) {
+	db := redis.NewClusterClient(&redis.ClusterOptions{
+		Addrs:              cfg.Addrs,
+		Username:           cfg.Username,
+		Password:           cfg.Password,
+		NewClient:          cfg.NewClient,
+		ClusterSlots:       cfg.ClusterSlots,
+		RouteByLatency:     cfg.RouteByLatency,
+		RouteRandomly:      cfg.RouteRandomly,
+		MaxRedirects:       cfg.MaxRedirects,
+		MaxRetries:         cfg.MaxRetries,
+		MinRetryBackoff:    cfg.MinRetryBackoff,
+		MaxRetryBackoff:    cfg.MaxRetryBackoff,
+		DialTimeout:        cfg.DialTimeout,
+		ReadTimeout:        cfg.ReadTimeout,
+		WriteTimeout:       cfg.WriteTimeout,
+		PoolSize:           cfg.PoolSize,
+		MinIdleConns:       cfg.MinIdleConns,
+		MaxConnAge:         cfg.MaxConnAge,
+		PoolTimeout:        cfg.PoolTimeout,
+		IdleTimeout:        cfg.IdleTimeout,
+		IdleCheckFrequency: cfg.IdleCheckFrequency,
+		TLSConfig:          cfg.TLSConfig,
+	})
+
+	if err := db.Ping(context.Background()).Err(); err != nil {
+		return nil, errRedisConnection(err)
+	}
+
+	p := &Provider{
+		keyprefix: cfg.KeyPrefix,
+		db:        db,
 	}
 
 	return p, nil
@@ -52,7 +133,7 @@ func New(cfg Config) (*Provider, error) {
 
 func (p *Provider) getRedisSessionKey(sessionID []byte) string {
 	key := bytebufferpool.Get()
-	key.SetString(p.config.KeyPrefix)
+	key.SetString(p.keyprefix)
 	key.WriteString(":")
 	key.Write(sessionID)
