--- providers/redisfailover/README.md
@@ -1,8 +1,0 @@
-# Redis Fail Over
-
-Redis Fail Over (Sentinel) provider implementation.
-
-Better encoder:
-
-- Encode: `session.MSGPEncode`
-- Decode: `session.MSGPDecode`
